スクリプトのリファクタリング
