+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 data/*
 model.pth
+__pycache__
